requirements.txt
asgiref==3.3.1
Django==3.1.7
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-model-utils==4.1.1
fontawesome-free==5.15.3
Pillow==8.1.2
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
gunicorn==20.1.0